refactor(db): break local-seed.ts into modular lib files
Extracts the monolithic local-seed.ts into focused modules under
packages/db/src/local-seed-lib/ (orgs, events, positions, users,
sequences) so each domain area can be read and extended independently.
